CG4 1-10まで完成

+++ lib/engine/Input/Input.cpp
@@ -32,7 +32,7 @@ Input* Input::GetInstance()
 }
 
 #pragma region キーボード
-unsigned char Input::GetKeyTime(int keyName)
+unsigned char Input::GetKeyTime(BYTE keyName)
 {
 	if (PushKey(keyName))
 	{
