+++ lib/engine/Input/Input.h
@@ -22,7 +22,7 @@ public:
 	void Initialize();
 	void Update();
 
-	unsigned char GetKeyTime(int keyName);
+	unsigned char GetKeyTime(BYTE keyName);
 	bool PushKey(BYTE keyNum);
 	bool ReleaseKey(BYTE keyNum);
 	bool TriggerPushKey(BYTE keyNum);
